refactor: выделила utils в сегмент приложения
