+++ app/presentation/telegram_bot_controller.py
@@ -4,7 +4,7 @@ import logging
 from aiogram import Bot, Dispatcher, Router, types
 from aiogram.filters import Command
 from app.presentation.command_dispatcher import ReminderDispatcher
-from app.schedulers.reminder_scheduler import ReminderScheduler
+from app.application.services.reminder_scheduler import ReminderScheduler
 
 logger = logging.getLogger(__name__)
