+++ pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
     "apscheduler>=3.11.2",
     "asyncpg>=0.31.0",
     "dataclasses>=0.8",
+    "freezegun>=1.5.5",
     "pytest>=9.0.2",
     "pytest-asyncio>=1.3.0",
     "python-dotenv>=1.2.2",
